Moved {Issue99, JsonForms}Spec to test root folder
